requirements.txt
gradio
ollama
pydub
openai-whisper